backend/requirements.dev.txt
# Minimal development requirements for InfoSentinel Backend
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Security
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.0.1

# Async and WebSocket
eventlet==0.33.3
python-socketio==5.8.0

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
click==8.1.7
markupsafe==2.1.3

# MongoDB
pymongo==4.5.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
requests-html==0.10.0

# Development tools
pytest==7.4.2
pytest-flask==1.2.0